telemetry: include CI events
Fixes #1205
